app/requirements.txt
ollama
yfinance
schedule
pytz
sqlalchemy>=2.0.0
finnhub-python
python-dotenv
